% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{medium2jekyll}
\alias{medium2jekyll}
\title{Runs Full Process of Porting Medium Blog}
\usage{
medium2jekyll(url, where_to = "~/Desktop/blog_transfer")
}
\arguments{
\item{url}{The Medium url or Medium @handle to port}

\item{where_to}{Folder location for ported content}
}
\description{
Assists in Migrating from a Medium Blog to a Jekyll-based Blog, e.g., Github.io.
}
\examples{
\dontrun{
medium2jekyll("@fun_times")
}
}

