% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{new_post}
\alias{new_post}
\title{Creates a New Post in Jekyll-based Markdown}
\usage{
new_post(title, subfolder = "./_posts/", date = as.character(Sys.Date()))
}
\arguments{
\item{title}{Title of new post}

\item{subfolder}{where file is saved}

\item{date}{Date, default is Sys.Date}
}
\description{
Creates a New Post in Jekyll-based Markdown
}
\examples{
\dontrun{
new_post("the end of the year", date = "2016-12-31")
}
}

