% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{new_post}
\alias{new_post}
\title{Creates a New Post in Jekyll-based Markdown}
\usage{
new_post(title, titlecase = TRUE, date = as.character(Sys.Date() + 1))
}
\arguments{
\item{title}{Title of new post}

\item{titlecase}{Whether title should be in titlecase}

\item{date}{Date, default is Sys.Date + 1}
}
\description{
Creates a New Post in Jekyll-based Markdown
}
\examples{
\dontrun{
new_post("the end of the year", date = "2016-12-31")
}
}

